Pre-api refactoring.
